Made minor tweaks (with Randall's help) to the python code.
Created (bash) wrapper script which is invoked from drupal
and creates the python environnment and calls the python
script which checks the data deposition area for the user.

Created the python script which formats the output of the
tests and outputs HTML which eventually gets displayed
via drupal for the ccfa user.